+++ lib/features/qr_scanner/screens/qr_scanner_with_image_screen.dart
@@ -645,7 +645,7 @@ class _QrScannerWithImageScreenState extends State<QrScannerWithImageScreen> {
 
       if (result != null) {
         logInfo(
-          'Code decoded [${result.formatName}]: ${result.text}',
+          'Code decoded [${result.formatName}]',
           tag: _logTag,
         );
       } else {